VoidKeepInv 1.1
This version introduces new features!

Players can be notified of items being restored by this plugin if the administrator chooses to do so.

Administrators can also decide if players should regain their XP after a void death.

Players can be told that XP cannot be restored due to settings set by their administrator, if the administrator chooses to do so.